@@ -3,13 +3,13 @@ module github.com/buildpacks/imgutil
 require (
 	github.com/docker/docker v1.4.2-0.20190924003213-a8608b5b67c7
 	github.com/docker/go-connections v0.4.0
-	github.com/google/go-cmp v0.3.0
+	github.com/google/go-cmp v0.4.0
 	github.com/google/go-containerregistry v0.0.0-20200311163244-4b1985e5ea21
-	github.com/pkg/errors v0.8.1
-	github.com/sclevine/spec v1.2.0
+	github.com/pkg/errors v0.9.1
+	github.com/sclevine/spec v1.4.0
 	golang.org/x/net v0.0.0-20200226121028-0de0cce0169b // indirect
 	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e // indirect
 	gopkg.in/yaml.v2 v2.2.8 // indirect
 )
 
-go 1.13
+go 1.14
