@@ -24,7 +24,9 @@ import (
 	"github.com/buildpacks/pack/internal/build"
 	"github.com/buildpacks/pack/internal/builder"
 	"github.com/buildpacks/pack/internal/buildpack"
+	"github.com/buildpacks/pack/internal/buildpackage"
 	"github.com/buildpacks/pack/internal/dist"
+	"github.com/buildpacks/pack/internal/layer"
 	"github.com/buildpacks/pack/internal/paths"
 	"github.com/buildpacks/pack/internal/stack"
 	"github.com/buildpacks/pack/internal/stringset"
@@ -36,8 +38,9 @@ type Lifecycle interface {
 }
 
 type BuildOptions struct {
-	Image              string              // required
-	Builder            string              // required
+	Image              string // required
+	Builder            string // required
+	Registry           string
 	AppPath            string              // defaults to current working directory
 	RunImage           string              // defaults to the best mirror from the builder metadata or AdditionalMirrors
 	AdditionalMirrors  map[string][]string // only considered if RunImage is not provided
@@ -102,7 +105,7 @@ func (c *Client) Build(ctx context.Context, opts BuildOptions) error {
 		return err
 	}
 
-	fetchedBPs, order, err := c.processBuildpacks(ctx, bldr.Buildpacks(), bldr.Order(), opts.Buildpacks, opts.NoPull, opts.Publish)
+	fetchedBPs, order, err := c.processBuildpacks(ctx, bldr.Image(), bldr.Buildpacks(), bldr.Order(), opts.Buildpacks, opts.NoPull, opts.Publish, opts.Registry)
 	if err != nil {
 		return err
 	}
@@ -394,7 +397,7 @@ func (c *Client) processProxyConfig(config *ProxyConfig) ProxyConfig {
 // 	----------
 // 	- group:
 //		- A
-func (c *Client) processBuildpacks(ctx context.Context, builderBPs []dist.BuildpackInfo, builderOrder dist.Order, declaredBPs []string, noPull bool, publish bool) (fetchedBPs []dist.Buildpack, order dist.Order, err error) {
+func (c *Client) processBuildpacks(ctx context.Context, builderImage imgutil.Image, builderBPs []dist.BuildpackInfo, builderOrder dist.Order, declaredBPs []string, noPull bool, publish bool, registry string) (fetchedBPs []dist.Buildpack, order dist.Order, err error) {
 	order = dist.Order{{Group: []dist.BuildpackRef{}}}
 	for _, bp := range declaredBPs {
 		locatorType, err := buildpack.GetLocatorType(bp, builderBPs)
@@ -437,17 +440,54 @@ func (c *Client) processBuildpacks(ctx context.Context, builderBPs []dist.Buildp
 				return fetchedBPs, order, errors.Wrapf(err, "downloading buildpack from %s", style.Symbol(bp))
 			}
 
-			fetchedBP, err := dist.BuildpackFromRootBlob(blob)
+			var mainBP dist.Buildpack
+			var dependencyBPs []dist.Buildpack
+
+			isOCILayout, err := buildpackage.IsOCILayoutBlob(blob)
 			if err != nil {
-				return fetchedBPs, order, errors.Wrapf(err, "creating buildpack from %s", style.Symbol(bp))
+				return fetchedBPs, order, errors.Wrapf(err, "checking format")
 			}
 
-			fetchedBPs = append(fetchedBPs, fetchedBP)
-			order = appendBuildpackToOrder(order, fetchedBP.Descriptor().Info)
+			if isOCILayout {
+				mainBP, dependencyBPs, err = buildpackage.BuildpacksFromOCILayoutBlob(blob)
+				if err != nil {
+					return fetchedBPs, order, errors.Wrapf(err, "extracting buildpacks from %s", style.Symbol(bp))
+				}
+			} else {
+				layerWriterFactory, err := layer.NewWriterFactory(builderImage)
+				if err != nil {
+					return fetchedBPs, order, errors.Wrapf(err, "get tar writer factory for image %s", style.Symbol(builderImage.Name()))
+				}
+				mainBP, err = dist.BuildpackFromRootBlob(blob, layerWriterFactory)
+				if err != nil {
+					return fetchedBPs, order, errors.Wrapf(err, "creating buildpack from %s", style.Symbol(bp))
+				}
+			}
+
+			fetchedBPs = append(append(fetchedBPs, mainBP), dependencyBPs...)
+			order = appendBuildpackToOrder(order, mainBP.Descriptor().Info)
 		case buildpack.PackageLocator:
 			mainBP, depBPs, err := extractPackagedBuildpacks(ctx, bp, c.imageFetcher, publish, noPull)
 			if err != nil {
 				return fetchedBPs, order, errors.Wrapf(err, "creating from buildpackage %s", style.Symbol(bp))
+			}
+
+			fetchedBPs = append(append(fetchedBPs, mainBP), depBPs...)
+			order = appendBuildpackToOrder(order, mainBP.Descriptor().Info)
+		case buildpack.RegistryLocator:
+			registryCache, err := c.getRegistry(c.logger, registry)
+			if err != nil {
+				return fetchedBPs, order, errors.Wrapf(err, "invalid registry '%s'", registry)
+			}
+
+			registryBp, err := registryCache.LocateBuildpack(bp)
+			if err != nil {
+				return fetchedBPs, order, errors.Wrapf(err, "locating in registry %s", style.Symbol(bp))
+			}
+
+			mainBP, depBPs, err := extractPackagedBuildpacks(ctx, registryBp.Address, c.imageFetcher, publish, noPull)
+			if err != nil {
+				return fetchedBPs, order, errors.Wrapf(err, "extracting from registry %s", style.Symbol(bp))
 			}
 
 			fetchedBPs = append(append(fetchedBPs, mainBP), depBPs...)
