@@ -2,11 +2,11 @@ package buildpackage
 
 import (
 	"path/filepath"
-	"strings"
 
 	"github.com/BurntSushi/toml"
 	"github.com/pkg/errors"
 
+	"github.com/buildpacks/pack/internal/config"
 	"github.com/buildpacks/pack/internal/dist"
 	"github.com/buildpacks/pack/internal/paths"
 	"github.com/buildpacks/pack/internal/style"
@@ -30,72 +30,52 @@ type ConfigReader struct{}
 // Read reads and validates a buildpackage configuration from the file path provided and returns the
 // configuration and any error that occurred during reading or validation.
 func (r *ConfigReader) Read(path string) (Config, error) {
-	config := Config{}
+	packageConfig := Config{}
 
-	tomlMetadata, err := toml.DecodeFile(path, &config)
+	tomlMetadata, err := toml.DecodeFile(path, &packageConfig)
 	if err != nil {
-		return config, errors.Wrap(err, "decoding toml")
+		return packageConfig, errors.Wrap(err, "decoding toml")
 	}
 
 	undecodedKeys := tomlMetadata.Undecoded()
 	if len(undecodedKeys) > 0 {
-		unusedKeys := map[string]interface{}{}
-		for _, key := range undecodedKeys {
-			keyName := key.String()
+		unknownElementsMsg := config.FormatUndecodedKeys(undecodedKeys)
 
-			parent := strings.Split(keyName, ".")[0]
-
-			if _, ok := unusedKeys[parent]; !ok {
-				unusedKeys[keyName] = nil
-			}
-		}
-
-		var errorKeys []string
-		for errorKey := range unusedKeys {
-			errorKeys = append(errorKeys, style.Symbol(errorKey))
-		}
-
-		pluralizedElement := "element"
-		if len(errorKeys) > 1 {
-			pluralizedElement += "s"
-		}
-
-		return config, errors.Errorf("unknown configuration %s %s in %s",
-			pluralizedElement,
-			strings.Join(errorKeys, ", "),
+		return packageConfig, errors.Errorf("%s in %s",
+			unknownElementsMsg,
 			style.Symbol(path),
 		)
 	}
 
-	if config.Buildpack.URI == "" {
-		return config, errors.Errorf("missing %s configuration", style.Symbol("buildpack.uri"))
+	if packageConfig.Buildpack.URI == "" {
+		return packageConfig, errors.Errorf("missing %s configuration", style.Symbol("buildpack.uri"))
 	}
 
 	configDir, err := filepath.Abs(filepath.Dir(path))
 	if err != nil {
-		return config, err
+		return packageConfig, err
 	}
 
-	absPath, err := paths.ToAbsolute(config.Buildpack.URI, configDir)
+	absPath, err := paths.ToAbsolute(packageConfig.Buildpack.URI, configDir)
 	if err != nil {
-		return config, errors.Wrapf(err, "getting absolute path for %s", style.Symbol(config.Buildpack.URI))
+		return packageConfig, errors.Wrapf(err, "getting absolute path for %s", style.Symbol(packageConfig.Buildpack.URI))
 	}
-	config.Buildpack.URI = absPath
+	packageConfig.Buildpack.URI = absPath
 
-	for i := range config.Dependencies {
-		uri := config.Dependencies[i].URI
+	for i := range packageConfig.Dependencies {
+		uri := packageConfig.Dependencies[i].URI
 		if uri != "" {
 			absPath, err := paths.ToAbsolute(uri, configDir)
 			if err != nil {
-				return config, errors.Wrapf(err, "getting absolute path for %s", style.Symbol(uri))
+				return packageConfig, errors.Wrapf(err, "getting absolute path for %s", style.Symbol(uri))
 			}
 
-			config.Dependencies[i].URI = absPath
+			packageConfig.Dependencies[i].URI = absPath
 		}
 
-		dep := config.Dependencies[i]
+		dep := packageConfig.Dependencies[i]
 		if dep.URI != "" && dep.ImageName != "" {
-			return config, errors.Errorf(
+			return packageConfig, errors.Errorf(
 				"dependency configured with both %s and %s",
 				style.Symbol("uri"),
 				style.Symbol("image"),
@@ -103,5 +83,5 @@ func (r *ConfigReader) Read(path string) (Config, error) {
 		}
 	}
 
-	return config, nil
+	return packageConfig, nil
 }
