@@ -28,7 +28,6 @@ func main() {
 		logger.Error(err.Error())
 		os.Exit(1)
 	}
-
 	rootCmd := &cobra.Command{
 		Use: "pack",
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {
@@ -63,7 +62,7 @@ func main() {
 	rootCmd.AddCommand(commands.Rebase(logger, cfg, &packClient))
 	rootCmd.AddCommand(commands.InspectImage(logger, &cfg, &packClient))
 
-	rootCmd.AddCommand(commands.CreateBuilder(logger, &packClient))
+	rootCmd.AddCommand(commands.CreateBuilder(logger, cfg, &packClient))
 	rootCmd.AddCommand(commands.PackageBuildpack(logger, &packClient, buildpackage.NewConfigReader()))
 	rootCmd.AddCommand(commands.SetRunImagesMirrors(logger, cfg))
 	rootCmd.AddCommand(commands.InspectBuilder(logger, cfg, &packClient))
