@@ -8,44 +8,120 @@ import (
 	"github.com/buildpacks/imgutil"
 	"github.com/pkg/errors"
 
+	"github.com/buildpacks/pack/internal/buildpack"
+	"github.com/buildpacks/pack/internal/buildpackage"
+
 	pubbldr "github.com/buildpacks/pack/builder"
 	"github.com/buildpacks/pack/internal/builder"
 	"github.com/buildpacks/pack/internal/dist"
 	"github.com/buildpacks/pack/internal/image"
+	"github.com/buildpacks/pack/internal/layer"
 	"github.com/buildpacks/pack/internal/style"
 )
 
+// CreateBuilderOptions are options passed to CreateBuilder
 type CreateBuilderOptions struct {
 	BuilderName string
 	Config      pubbldr.Config
 	Publish     bool
 	NoPull      bool
+	Registry    string
 }
 
+// CreateBuilder creates a builder
 func (c *Client) CreateBuilder(ctx context.Context, opts CreateBuilderOptions) error {
-	if err := validateBuilderConfig(opts.Config); err != nil {
+	if err := c.validateConfig(ctx, opts); err != nil {
+		return err
+	}
+
+	builder, err := c.createBaseBuilder(ctx, opts)
+	if err != nil {
+		return errors.Wrap(err, "failed to create builder")
+	}
+
+	if err := c.addBuildpacksToBuilder(ctx, opts, builder); err != nil {
+		return errors.Wrap(err, "failed to add buildpacks to builder")
+	}
+
+	builder.SetOrder(opts.Config.Order)
+	builder.SetStack(opts.Config.Stack)
+
+	return builder.Save(c.logger)
+}
+
+func (c *Client) validateConfig(ctx context.Context, opts CreateBuilderOptions) error {
+	if err := opts.Config.Validate(); err != nil {
 		return errors.Wrap(err, "invalid builder config")
 	}
 
 	if err := c.validateRunImageConfig(ctx, opts); err != nil {
-		return err
+		return errors.Wrap(err, "invalid run image config")
 	}
 
+	return nil
+}
+
+func (c *Client) validateRunImageConfig(ctx context.Context, opts CreateBuilderOptions) error {
+	var runImages []imgutil.Image
+	for _, i := range append([]string{opts.Config.Stack.RunImage}, opts.Config.Stack.RunImageMirrors...) {
+		if !opts.Publish {
+			img, err := c.imageFetcher.Fetch(ctx, i, true, false)
+			if err != nil {
+				if errors.Cause(err) != image.ErrNotFound {
+					return errors.Wrap(err, "failed to fetch image")
+				}
+			} else {
+				runImages = append(runImages, img)
+				continue
+			}
+		}
+
+		img, err := c.imageFetcher.Fetch(ctx, i, false, false)
+		if err != nil {
+			if errors.Cause(err) != image.ErrNotFound {
+				return errors.Wrap(err, "failed to fetch image")
+			}
+			c.logger.Warnf("run image %s is not accessible", style.Symbol(i))
+		} else {
+			runImages = append(runImages, img)
+		}
+	}
+
+	for _, img := range runImages {
+		stackID, err := img.Label("io.buildpacks.stack.id")
+		if err != nil {
+			return errors.Wrap(err, "failed to label image")
+		}
+
+		if stackID != opts.Config.Stack.ID {
+			return fmt.Errorf(
+				"stack %s from builder config is incompatible with stack %s from run image %s",
+				style.Symbol(opts.Config.Stack.ID),
+				style.Symbol(stackID),
+				style.Symbol(img.Name()),
+			)
+		}
+	}
+
+	return nil
+}
+
+func (c *Client) createBaseBuilder(ctx context.Context, opts CreateBuilderOptions) (*builder.Builder, error) {
 	baseImage, err := c.imageFetcher.Fetch(ctx, opts.Config.Stack.BuildImage, !opts.Publish, !opts.NoPull)
 	if err != nil {
-		return errors.Wrap(err, "fetch build image")
+		return nil, errors.Wrap(err, "fetch build image")
 	}
 
 	c.logger.Debugf("Creating builder %s from build-image %s", style.Symbol(opts.BuilderName), style.Symbol(baseImage.Name()))
 	bldr, err := builder.New(baseImage, opts.BuilderName)
 	if err != nil {
-		return errors.Wrap(err, "invalid build-image")
+		return nil, errors.Wrap(err, "invalid build-image")
 	}
 
 	bldr.SetDescription(opts.Config.Description)
 
 	if bldr.StackID != opts.Config.Stack.ID {
-		return fmt.Errorf(
+		return nil, fmt.Errorf(
 			"stack %s from builder config is incompatible with stack %s from build image",
 			style.Symbol(opts.Config.Stack.ID),
 			style.Symbol(bldr.StackID),
@@ -54,74 +130,12 @@ func (c *Client) CreateBuilder(ctx context.Context, opts CreateBuilderOptions) e
 
 	lifecycle, err := c.fetchLifecycle(ctx, opts.Config.Lifecycle)
 	if err != nil {
-		return errors.Wrap(err, "fetch lifecycle")
+		return nil, errors.Wrap(err, "fetch lifecycle")
 	}
 
-	if err := bldr.SetLifecycle(lifecycle); err != nil {
-		return errors.Wrap(err, "setting lifecycle")
-	}
+	bldr.SetLifecycle(lifecycle)
 
-	for _, b := range opts.Config.Buildpacks.Buildpacks() {
-		err := ensureBPSupport(b.URI)
-		if err != nil {
-			return err
-		}
-
-		blob, err := c.downloader.Download(ctx, b.URI)
-		if err != nil {
-			return errors.Wrapf(err, "downloading buildpack from %s", style.Symbol(b.URI))
-		}
-
-		fetchedBp, err := dist.BuildpackFromRootBlob(blob)
-		if err != nil {
-			return errors.Wrapf(err, "creating buildpack from %s", style.Symbol(b.URI))
-		}
-
-		err = validateBuildpack(fetchedBp, b.URI, b.ID, b.Version)
-		if err != nil {
-			return errors.Wrap(err, "invalid buildpack")
-		}
-
-		bldr.AddBuildpack(fetchedBp)
-	}
-
-	for _, pkg := range opts.Config.Buildpacks.Packages() {
-		mainBP, depBPs, err := extractPackagedBuildpacks(ctx, pkg.ImageName, c.imageFetcher, opts.Publish, opts.NoPull)
-		if err != nil {
-			return err
-		}
-
-		for _, bp := range append([]dist.Buildpack{mainBP}, depBPs...) {
-			bldr.AddBuildpack(bp)
-		}
-	}
-
-	bldr.SetOrder(opts.Config.Order)
-	bldr.SetStack(opts.Config.Stack)
-
-	return bldr.Save(c.logger)
-}
-
-func validateBuildpack(bp dist.Buildpack, source, expectedID, expectedBPVersion string) error {
-	if expectedID != "" && bp.Descriptor().Info.ID != expectedID {
-		return fmt.Errorf(
-			"buildpack from URI %s has ID %s which does not match ID %s from builder config",
-			style.Symbol(source),
-			style.Symbol(bp.Descriptor().Info.ID),
-			style.Symbol(expectedID),
-		)
-	}
-
-	if expectedBPVersion != "" && bp.Descriptor().Info.Version != expectedBPVersion {
-		return fmt.Errorf(
-			"buildpack from URI %s has version %s which does not match version %s from builder config",
-			style.Symbol(source),
-			style.Symbol(bp.Descriptor().Info.Version),
-			style.Symbol(expectedBPVersion),
-		)
-	}
-
-	return nil
+	return bldr, nil
 }
 
 func (c *Client) fetchLifecycle(ctx context.Context, config pubbldr.LifecycleConfig) (builder.Lifecycle, error) {
@@ -160,67 +174,119 @@ func (c *Client) fetchLifecycle(ctx context.Context, config pubbldr.LifecycleCon
 	return lifecycle, nil
 }
 
-func uriFromLifecycleVersion(version semver.Version) string {
-	return fmt.Sprintf("https://github.com/buildpacks/lifecycle/releases/download/v%s/lifecycle-v%s+linux.x86-64.tgz", version.String(), version.String())
-}
-
-func validateBuilderConfig(conf pubbldr.Config) error {
-	if conf.Stack.ID == "" {
-		return errors.New("stack.id is required")
-	}
-
-	if conf.Stack.BuildImage == "" {
-		return errors.New("stack.build-image is required")
-	}
-
-	if conf.Stack.RunImage == "" {
-		return errors.New("stack.run-image is required")
-	}
-
-	return nil
-}
-
-func (c *Client) validateRunImageConfig(ctx context.Context, opts CreateBuilderOptions) error {
-	var runImages []imgutil.Image
-	for _, i := range append([]string{opts.Config.Stack.RunImage}, opts.Config.Stack.RunImageMirrors...) {
-		if !opts.Publish {
-			img, err := c.imageFetcher.Fetch(ctx, i, true, false)
-			if err != nil {
-				if errors.Cause(err) != image.ErrNotFound {
-					return err
-				}
-			} else {
-				runImages = append(runImages, img)
-				continue
-			}
+func (c *Client) addBuildpacksToBuilder(ctx context.Context, opts CreateBuilderOptions, bldr *builder.Builder) error {
+	for _, b := range opts.Config.Buildpacks {
+		locator := b.URI
+		if locator == "" && b.ImageName != "" {
+			locator = b.ImageName
 		}
 
-		img, err := c.imageFetcher.Fetch(ctx, i, false, false)
-		if err != nil {
-			if errors.Cause(err) != image.ErrNotFound {
-				return err
-			}
-			c.logger.Warnf("run image %s is not accessible", style.Symbol(i))
-		} else {
-			runImages = append(runImages, img)
-		}
-	}
-
-	for _, img := range runImages {
-		stackID, err := img.Label("io.buildpacks.stack.id")
+		locatorType, err := buildpack.GetLocatorType(locator, []dist.BuildpackInfo{})
 		if err != nil {
 			return err
 		}
 
-		if stackID != opts.Config.Stack.ID {
-			return fmt.Errorf(
-				"stack %s from builder config is incompatible with stack %s from run image %s",
-				style.Symbol(opts.Config.Stack.ID),
-				style.Symbol(stackID),
-				style.Symbol(img.Name()),
-			)
+		var bps []dist.Buildpack
+		switch locatorType {
+		case buildpack.RegistryLocator:
+			registryCache, err := c.getRegistry(c.logger, opts.Registry)
+			if err != nil {
+				return errors.Wrapf(err, "invalid registry '%s'", opts.Registry)
+			}
+
+			registryBp, err := registryCache.LocateBuildpack(b.URI)
+			if err != nil {
+				return errors.Wrapf(err, "locating in registry %s", style.Symbol(b.URI))
+			}
+
+			mainBP, depBPs, err := extractPackagedBuildpacks(ctx, registryBp.Address, c.imageFetcher, opts.Publish, opts.NoPull)
+			if err != nil {
+				return errors.Wrapf(err, "extracting from registry %s", style.Symbol(b.URI))
+			}
+
+			bps = append([]dist.Buildpack{mainBP}, depBPs...)
+		case buildpack.PackageLocator:
+			mainBP, depBPs, err := extractPackagedBuildpacks(ctx, b.ImageName, c.imageFetcher, opts.Publish, opts.NoPull)
+			if err != nil {
+				return err
+			}
+
+			bps = append([]dist.Buildpack{mainBP}, depBPs...)
+		case buildpack.URILocator:
+			err := ensureBPSupport(b.URI)
+			if err != nil {
+				return err
+			}
+
+			blob, err := c.downloader.Download(ctx, b.URI)
+			if err != nil {
+				return errors.Wrapf(err, "downloading buildpack from %s", style.Symbol(b.URI))
+			}
+
+			isOCILayout, err := buildpackage.IsOCILayoutBlob(blob)
+			if err != nil {
+				return errors.Wrap(err, "inspecting buildpack blob")
+			}
+
+			if isOCILayout {
+				mainBP, depBPs, err := buildpackage.BuildpacksFromOCILayoutBlob(blob)
+				if err != nil {
+					return errors.Wrapf(err, "extracting buildpacks from %s", style.Symbol(b.ID))
+				}
+
+				bps = append([]dist.Buildpack{mainBP}, depBPs...)
+			} else {
+				layerWriterFactory, err := layer.NewWriterFactory(bldr.Image())
+				if err != nil {
+					return errors.Wrapf(err, "get tar writer factory for image %s", style.Symbol(bldr.Name()))
+				}
+
+				fetchedBp, err := dist.BuildpackFromRootBlob(blob, layerWriterFactory)
+				if err != nil {
+					return errors.Wrapf(err, "creating buildpack from %s", style.Symbol(b.URI))
+				}
+
+				err = validateBuildpack(fetchedBp, b.URI, b.ID, b.Version)
+				if err != nil {
+					return errors.Wrap(err, "invalid buildpack")
+				}
+
+				bps = []dist.Buildpack{fetchedBp}
+			}
+		default:
+			return fmt.Errorf("error reading %s: invalid locator: %s", locator, locatorType)
+		}
+
+		for _, bp := range bps {
+			bldr.AddBuildpack(bp)
 		}
 	}
 
 	return nil
+}
+
+func validateBuildpack(bp dist.Buildpack, source, expectedID, expectedBPVersion string) error {
+	if expectedID != "" && bp.Descriptor().Info.ID != expectedID {
+		return fmt.Errorf(
+			"buildpack from URI %s has ID %s which does not match ID %s from builder config",
+			style.Symbol(source),
+			style.Symbol(bp.Descriptor().Info.ID),
+			style.Symbol(expectedID),
+		)
+	}
+
+	if expectedBPVersion != "" && bp.Descriptor().Info.Version != expectedBPVersion {
+		return fmt.Errorf(
+			"buildpack from URI %s has version %s which does not match version %s from builder config",
+			style.Symbol(source),
+			style.Symbol(bp.Descriptor().Info.Version),
+			style.Symbol(expectedBPVersion),
+		)
+	}
+
+	return nil
+}
+
+func uriFromLifecycleVersion(version semver.Version) string {
+	return fmt.Sprintf("https://github.com/buildpacks/lifecycle/releases/download/v%s/lifecycle-v%s+linux.x86-64.tgz", version.String(), version.String())
 }
