@@ -42,35 +42,35 @@ func (t *TarBuilder) AddDir(path string, mode int64, modTime time.Time) {
 	})
 }
 
-func (t *TarBuilder) Reader() io.ReadCloser {
+func (t *TarBuilder) Reader(twf TarWriterFactory) io.ReadCloser {
 	pr, pw := io.Pipe()
 	go func() {
 		var err error
 		defer func() {
 			pw.CloseWithError(err)
 		}()
-		_, err = t.WriteTo(pw)
+		_, err = t.WriteTo(pw, twf)
 	}()
 
 	return pr
 }
 
-func (t *TarBuilder) WriteToPath(path string) error {
+func (t *TarBuilder) WriteToPath(path string, twf TarWriterFactory) error {
 	fh, err := os.Create(path)
 	if err != nil {
 		return errors.Wrapf(err, "create file for tar: %s", style.Symbol(path))
 	}
 	defer fh.Close()
 
-	_, err = t.WriteTo(fh)
+	_, err = t.WriteTo(fh, twf)
 	return err
 }
 
-func (t *TarBuilder) WriteTo(writer io.Writer) (int64, error) {
-	tw := tar.NewWriter(writer)
+func (t *TarBuilder) WriteTo(w io.Writer, twf TarWriterFactory) (int64, error) {
+	var written int64
+	tw := twf.NewWriter(w)
 	defer tw.Close()
 
-	var written int64
 	for _, f := range t.files {
 		if err := tw.WriteHeader(&tar.Header{
 			Typeflag: f.typeFlag,
