@@ -3,6 +3,8 @@ package commands
 import (
 	"fmt"
 
+	"github.com/buildpacks/pack/internal/config"
+
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
@@ -16,15 +18,27 @@ type CreateBuilderFlags struct {
 	BuilderTomlPath string
 	Publish         bool
 	NoPull          bool
+	Registry        string
 }
 
-func CreateBuilder(logger logging.Logger, client PackClient) *cobra.Command {
+func (c CreateBuilderFlags) validate() error {
+	if c.Publish && c.NoPull {
+		return errors.Errorf("The --publish and --no-pull flags cannot be used together. The --publish flag requires the use of remote images.")
+	}
+	return nil
+}
+
+func CreateBuilder(logger logging.Logger, cfg config.Config, client PackClient) *cobra.Command {
 	var flags CreateBuilderFlags
 	cmd := &cobra.Command{
 		Use:   "create-builder <image-name> --builder-config <builder-config-path>",
 		Args:  cobra.ExactArgs(1),
 		Short: "Create builder image",
 		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
+			if err := flags.validate(); err != nil {
+				return err
+			}
+
 			builderConfig, warns, err := builder.ReadConfig(flags.BuilderTomlPath)
 			if err != nil {
 				return errors.Wrap(err, "invalid builder toml")
@@ -39,6 +53,7 @@ func CreateBuilder(logger logging.Logger, client PackClient) *cobra.Command {
 				Config:      builderConfig,
 				Publish:     flags.Publish,
 				NoPull:      flags.NoPull,
+				Registry:    flags.Registry,
 			}); err != nil {
 				return err
 			}
@@ -49,6 +64,7 @@ func CreateBuilder(logger logging.Logger, client PackClient) *cobra.Command {
 	}
 	cmd.Flags().BoolVar(&flags.NoPull, "no-pull", false, "Skip pulling build image before use")
 	cmd.Flags().StringVarP(&flags.BuilderTomlPath, "builder-config", "b", "", "Path to builder TOML file (required)")
+	cmd.Flags().StringVarP(&flags.Registry, "buildpack-registry", "R", cfg.DefaultRegistry, "Buildpack Registry URL")
 	cmd.MarkFlagRequired("builder-config")
 	cmd.Flags().BoolVar(&flags.Publish, "publish", false, "Publish to registry")
 	AddHelpFlag(cmd, "create-builder")
