@@ -6,6 +6,7 @@ import (
 	"github.com/pkg/errors"
 
 	pubbldpkg "github.com/buildpacks/pack/buildpackage"
+	"github.com/buildpacks/pack/internal/archive"
 	"github.com/buildpacks/pack/internal/buildpackage"
 	"github.com/buildpacks/pack/internal/dist"
 	"github.com/buildpacks/pack/internal/style"
@@ -16,6 +17,7 @@ const (
 	FormatFile  = "file"
 )
 
+// PackageBuildpackOptions are configuration options and metadata you can pass into PackageBuildpack
 type PackageBuildpackOptions struct {
 	Name    string
 	Format  string
@@ -24,6 +26,7 @@ type PackageBuildpackOptions struct {
 	NoPull  bool
 }
 
+// PackageBuildpack packages buildpack(s) into an image or file
 func (c *Client) PackageBuildpack(ctx context.Context, opts PackageBuildpackOptions) error {
 	packageBuilder := buildpackage.NewBuilder(c.imageFactory)
 
@@ -41,7 +44,7 @@ func (c *Client) PackageBuildpack(ctx context.Context, opts PackageBuildpackOpti
 		return errors.Wrapf(err, "downloading buildpack from %s", style.Symbol(bpURI))
 	}
 
-	bp, err := dist.BuildpackFromRootBlob(blob)
+	bp, err := dist.BuildpackFromRootBlob(blob, archive.DefaultTarWriterFactory())
 	if err != nil {
 		return errors.Wrapf(err, "creating buildpack from %s", style.Symbol(bpURI))
 	}
@@ -55,7 +58,7 @@ func (c *Client) PackageBuildpack(ctx context.Context, opts PackageBuildpackOpti
 				return errors.Wrapf(err, "downloading buildpack from %s", style.Symbol(dep.URI))
 			}
 
-			depBP, err := dist.BuildpackFromRootBlob(blob)
+			depBP, err := dist.BuildpackFromRootBlob(blob, archive.DefaultTarWriterFactory())
 			if err != nil {
 				return errors.Wrapf(err, "creating buildpack from %s", style.Symbol(dep.URI))
 			}
