@@ -1,7 +1,6 @@
 package testhelpers
 
 import (
-	"bytes"
 	"context"
 	"encoding/base64"
 	"fmt"
@@ -17,9 +16,10 @@ import (
 	dockertypes "github.com/docker/docker/api/types"
 	dockercontainer "github.com/docker/docker/api/types/container"
 	"github.com/docker/go-connections/nat"
+	"golang.org/x/crypto/bcrypt"
 )
 
-var registryContainerName = "registry:2"
+var registryContainerName = "cnbs/registry:2"
 
 type TestRegistryConfig struct {
 	runRegistryName string
@@ -48,6 +48,23 @@ func RunRegistry(t *testing.T) *TestRegistryConfig {
 		password:        password,
 	}
 
+	// ensure registry is ready
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	for {
+		_, err := registryConfig.RegistryCatalog()
+		if err == nil {
+			break
+		}
+
+		err = ctx.Err()
+		if err != nil {
+			t.Fatal("registry not ready:", err.Error())
+		}
+
+		time.Sleep(500 * time.Microsecond)
+	}
+
 	return registryConfig
 }
 
@@ -72,7 +89,8 @@ func startRegistry(t *testing.T, runRegistryName, username, password string) str
 	AssertNil(t, PullImageWithAuth(dockerCli(t), registryContainerName, ""))
 	ctx := context.Background()
 
-	htpasswdTar := generateHtpasswd(ctx, t, username, password)
+	htpasswdTar := generateHtpasswd(t, username, password)
+	defer htpasswdTar.Close()
 
 	ctr, err := dockerCli(t).ContainerCreate(ctx, &dockercontainer.Config{
 		Image:  registryContainerName,
@@ -106,22 +124,11 @@ func startRegistry(t *testing.T, runRegistryName, username, password string) str
 	return runRegistryPort
 }
 
-func generateHtpasswd(ctx context.Context, t *testing.T, username string, password string) io.Reader {
-	//https://docs.docker.com/registry/deploying/#restricting-access
-	htpasswdCtr, err := dockerCli(t).ContainerCreate(ctx, &dockercontainer.Config{
-		Image:      registryContainerName,
-		Entrypoint: []string{"htpasswd", "-Bbn", username, password},
-	}, &dockercontainer.HostConfig{
-		AutoRemove: true,
-	}, nil, "")
-	AssertNil(t, err)
-
-	var b bytes.Buffer
-	err = RunContainer(ctx, dockerCli(t), htpasswdCtr.ID, &b, &b)
-	AssertNil(t, err)
-	reader, err := archive.CreateSingleFileTarReader("/registry_test_htpasswd", b.String())
-	AssertNil(t, err)
-
+func generateHtpasswd(t *testing.T, username string, password string) io.ReadCloser {
+	// https://docs.docker.com/registry/deploying/#restricting-access
+	// HTPASSWD format: https://github.com/foomo/htpasswd/blob/e3a90e78da9cff06a83a78861847aa9092cbebdd/hashing.go#L23
+	passwordBytes, _ := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
+	reader := archive.CreateSingleFileTarReader("/registry_test_htpasswd", username+":"+string(passwordBytes))
 	return reader
 }
 
